+++ src/factory/error/notFound.ts
@@ -6,7 +6,7 @@ import { WaiterError } from './waiter';
 /**
  * NotFoundError
  */
-export default class NotFoundError extends WaiterError {
+export class NotFoundError extends WaiterError {
     public readonly entityName: string;
 
     constructor(entityName: string, message?: string) {