+++ src/factory/error/unauthorized.spec.ts
@@ -3,21 +3,21 @@
  */
 import * as assert from 'assert';
 
-import { ServiceUnavailableError } from './serviceUnavailable';
+import { UnauthorizedError } from './unauthorized';
 
-describe('new ServiceUnavailableError()', () => {
+describe('new UnauthorizedError()', () => {
     it('正しくインスタンス化できる', () => {
         const message = 'test message';
-        const error = new ServiceUnavailableError(message);
-        assert(error instanceof ServiceUnavailableError);
+        const error = new UnauthorizedError(message);
+        assert(error instanceof UnauthorizedError);
         assert.equal(error.message, message);
         assert.equal(error.name, 'WaiterError');
         assert.equal(typeof error.stack, 'string');
     });
 
     it('メッセージを指定しなくても、正しくインスタンス化できる', () => {
-        const error = new ServiceUnavailableError();
-        assert(error instanceof ServiceUnavailableError);
+        const error = new UnauthorizedError();
+        assert(error instanceof UnauthorizedError);
         assert.equal(error.name, 'WaiterError');
         assert.equal(typeof error.message, 'string');
         assert.equal(typeof error.stack, 'string');