+++ src/factory/errorCode.ts
@@ -9,6 +9,7 @@ enum ErrorCode {
     NotFound = 'NotFound',
     NotImplemented = 'NotImplemented',
     ServiceUnavailable = 'ServiceUnavailable',
-    RateLimitExceeded = 'RateLimitExceeded'
+    RateLimitExceeded = 'RateLimitExceeded',
+    Unauthorized = 'Unauthorized'
 }
 export default ErrorCode;