+++ src/factory/errors.ts
@@ -1,19 +1,22 @@
 /**
  * errors
  */
-import ArgumentError from './error/argument';
-import ArgumentNullError from './error/argumentNull';
-import ForbiddenError from './error/forbidden';
-import NotFoundError from './error/notFound';
-import RateLimitExceededError from './error/rateLimitExceeded';
-import ServiceUnavailableError from './error/serviceUnavailable';
-import { WaiterError } from './error/waiter';
+import { ArgumentError as Argument } from './error/argument';
+import { ArgumentNullError as ArgumentNull } from './error/argumentNull';
+import { ForbiddenError as Forbidden } from './error/forbidden';
+import { NotFoundError as NotFound } from './error/notFound';
+import { RateLimitExceededError as RateLimitExceeded } from './error/rateLimitExceeded';
+import { ServiceUnavailableError as ServiceUnavailable } from './error/serviceUnavailable';
+import { UnauthorizedError as Unauthorized } from './error/unauthorized';
+import { WaiterError as Waiter } from './error/waiter';
+
 export {
-    ArgumentError as Argument,
-    ArgumentNullError as ArgumentNull,
-    ForbiddenError as Forbidden,
-    NotFoundError as NotFound,
-    RateLimitExceededError as RateLimitExceeded,
-    ServiceUnavailableError as ServiceUnavailable,
-    WaiterError as Waiter
+    Argument,
+    ArgumentNull,
+    Forbidden,
+    NotFound,
+    RateLimitExceeded,
+    ServiceUnavailable,
+    Unauthorized,
+    Waiter
 };