+++ src/factory/error/waiter.ts
@@ -1,4 +1,4 @@
-import ErrorCode from '../errorCode';
+import { ErrorCode } from '../errorCode';
 
 /**
  * WaiterError