+++ src/index.ts
@@ -2,15 +2,19 @@
  * factory index
  */
 import * as client from './factory/client';
-import ErrorCode from './factory/errorCode';
+import { ErrorCode } from './factory/errorCode';
 import * as errors from './factory/errors';
 import * as passport from './factory/passport';
 import * as project from './factory/project';
 import * as rule from './factory/rule';
+import * as ruleSet from './factory/ruleSet';
 
-export import client = client;
-export import errorCode = ErrorCode;
-export import errors = errors;
-export import passport = passport;
-export import project = project;
-export import rule = rule;
+export {
+    client,
+    ErrorCode as errorCode,
+    errors,
+    passport,
+    project,
+    rule,
+    ruleSet
+};